+++ binary/bibmaptitlecases.py
@@ -13,43 +13,58 @@ CaseProtect=['AIAA','IEEE','AAAI',"GECCO\'18","ACML","SAR",'BVR','UAV','UCAV','U
 
 '''
 sourcemaps=[
-	[#map1:将设置title为sentencecase
-		{"fieldsource":"title","final":True},#
-		{"fieldset":"title","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
-	],
+    [#map1:将设置title为sentencecase
+        {"fieldsource":"title","final":True},#
+        {"fieldset":"title","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
+    ],
     [#map2:将设置booktitle为titlecase
-		{"fieldsource":"booktitle","final":True},#
-		{"fieldset":"booktitle","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
-	],
-	[#map3:将设置journal为titlecase
-		{"fieldsource":"journal","final":True},#
-		{"fieldset":"journal","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
-	],
-	[#map3:将设置journaltitle为titlecase
-		{"fieldsource":"journaltitle","final":True},#
-		{"fieldset":"journaltitle","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
-	]
+        {"fieldsource":"booktitle","final":True},#
+        {"fieldset":"booktitle","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
+    ],
+    [#map3:将设置journal为titlecase
+        {"fieldsource":"journal","final":True},#
+        {"fieldset":"journal","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
+    ],
+    [#map3:将设置journaltitle为titlecase
+        {"fieldsource":"journaltitle","final":True},#
+        {"fieldset":"journaltitle","origfieldval":True,"fieldfunction":'setuppercase',"overwrite":True}#step1
+    ]
 ]
 '''
 
 sourcemaps=[
-	[#map1:将设置title为sentencecase
-		{"fieldsource":"title","final":True},#
-		{"fieldset":"title","origfieldval":True,"fieldfunction":'setsentencecase',"overwrite":True}#step1
-	],
+    [#map1:将设置title为sentencecase
+        {"fieldsource":"title","final":True},#
+        {"fieldset":"title","origfieldval":True,"fieldfunction":'setsentencecase',"overwrite":True}#step1
+    ],
     [#map2:将设置booktitle为titlecase
-		{"fieldsource":"booktitle","final":True},#
-		{"fieldset":"booktitle","origfieldval":True,"fieldfunction":'settitlecasestd',"overwrite":True}#step1
-	],
-	[#map3:将设置journal为titlecase
-		{"fieldsource":"journal","final":True},#
-		{"fieldset":"journal","origfieldval":True,"fieldfunction":'settitlecasestd',"overwrite":True}#step1
-	],
-	[#map3:将设置journaltitle为titlecase
-		{"fieldsource":"journaltitle","final":True},#
-		{"fieldset":"journaltitle","origfieldval":True,"fieldfunction":'settitlecasestd',"overwrite":True}#step1
-	]
+        {"fieldsource":"booktitle","final":True},#
+        {"fieldset":"booktitle","origfieldval":True,"fieldfunction":'settitlecasestd',"overwrite":True}#step1
+    ],
+    [#map3:将设置journal为titlecase
+        {"fieldsource":"journal","final":True},#
+        {"fieldset":"journal","origfieldval":True,"fieldfunction":'settitlecasestd',"overwrite":True}#step1
+    ],
+    [#map3:将设置journaltitle为titlecase
+        {"fieldsource":"journaltitle","final":True},#
+        {"fieldset":"journaltitle","origfieldval":True,"fieldfunction":'settitlecasestd',"overwrite":True}#step1
+    ]
+    ]
+
+'''
+sourcemaps=[
+    [#map1:将设置title为sentencecase
+        {"fieldsource":"date","extract":r"(\d\d\d\d)","notinrange":['1995','2005'],"final":True},#
+        {"fieldsource":"title","final":True},#
+        {"fieldset":"title","origfieldval":True,"fieldfunction":'setsentencecase',"overwrite":True}#step1
+    ],
+    [#map2:将设置title为sentencecase
+        {"fieldsource":"entrykey","in":['southey2005bayes','Silver2017Mastering'],"final":True},#
+        {"fieldsource":"title","final":True},#
+        {"fieldset":"title","origfieldval":True,"fieldfunction":'setsentencecase',"overwrite":True}#step1
+    ]
 ]
+'''
 
 
 #默认不做修改